update new architecture, need to remove zoom
