+++ app.json
@@ -42,7 +42,7 @@
       "eas": {
         "projectId": "your-project-id"
       },
-      "API_ORIGIN": "https://8aebc705b28d.ngrok-free.app"
+      "API_ORIGIN": "https://7a49c7c18825.ngrok-free.app"
     },
     "owner": "your-expo-username"
   }
